Small graph on active contracts per month
