+++ package.json
@@ -66,6 +66,7 @@
     "react-scripts": "3.4.1",
     "react-select": "^3.1.0",
     "react-test-renderer": "^16.6.3",
+    "recharts":"1.8.5",
     "rollup": "^2.10.0"
   },
   "files": [
